docs(examples): swap oracle to median price feed + fix install/compile
Replace the multi-rpc-consensus-oracle example with a more useful
multi-source-price-oracle. Fetches ETH/BTC/SOL spot prices from three
keyless public exchanges (Coinbase, Kraken, Bitstamp) in parallel, takes
the median, rejects the result if the spread between min and max
exceeds maxSpreadBps (default 1%), and signs into a PriceOracle
registry. Median (rather than strict byte-equality) is the right
aggregation for live prices; one outage or one outlier doesn't break
the feed.

Includes a zero-dep `npm run test-medianizer` harness that runs the
fetch + aggregation logic locally without any chain or Lit involvement.
Verified live against the three APIs: ETH at 2bps spread, BTC at 3bps.

Also fixes a real install/compile blocker across all three examples:
@nomicfoundation/hardhat-toolbox@5.0.0 requires ethers v6, but Lit
ships ethers v5 inside the action runtime — and the off-chain scripts
were written for v5. Pinning hardhat-toolbox to ^2.0.2 (last line that
supports ethers v5) lets `npm install` resolve. Two deploy.js files
also had v6 API calls (`waitForDeployment`, `getAddress()`,
`ethers.parseUnits`) that would have crashed even with the right
toolbox — fixed those to v5 conventions to match the rest of the
off-chain stack. Cancun evmVersion set explicitly so OpenZeppelin v5's
`mcopy` use compiles.

Cross-references in docs section 9 and 10, examples/README.md,
compliance-transfer-gate README, and prediction-market-oracle README +
action all updated to point at the renamed example and explain the
median-vs-strict-agreement distinction.
