+++ examples/compliance-transfer-gate/hardhat.config.js
@@ -9,7 +9,12 @@ const BASE_SEPOLIA_RPC_URL =
 module.exports = {
   solidity: {
     version: "0.8.24",
-    settings: { optimizer: { enabled: true, runs: 200 } },
+    settings: {
+      optimizer: { enabled: true, runs: 200 },
+      // Required for OpenZeppelin v5 (uses Cancun's mcopy opcode).
+      // Base, Optimism, Arbitrum and Ethereum mainnet all support Cancun.
+      evmVersion: "cancun",
+    },
   },
   networks: {
     base: {
