+++ examples/compliance-transfer-gate/package.json
@@ -16,7 +16,7 @@
     "ethers": "^5.7.2"
   },
   "devDependencies": {
-    "@nomicfoundation/hardhat-toolbox": "^5.0.0",
+    "@nomicfoundation/hardhat-toolbox": "^2.0.2",
     "hardhat": "^2.22.10"
   }
 }
