--- examples/multi-rpc-consensus-oracle/.env.example
@@ -1,46 +1,0 @@
-# Lit Chipotle ---------------------------------------------------------------
-LIT_API_BASE=https://api.chipotle.litprotocol.com
-LIT_USAGE_API_KEY=
-
-# PKP used SOLELY as the encryption boundary for the three RPC URLs. Signs
-# nothing the registry cares about.
-DECRYPT_PKP_ADDRESS=
-
-# Wallet address derived from the action's IPFS CID. This is what the
-# deployed ConsensusOracle trusts as its signer. Editing the action source
-# produces a new CID and therefore a new address — old deployments stop
-# accepting signatures from the new action.
-ACTION_WALLET_ADDRESS=
-
-# RPC URLs (these embed provider API keys — do NOT commit) -------------------
-INFURA_URL=https://mainnet.infura.io/v3/...
-ALCHEMY_URL=https://eth-mainnet.g.alchemy.com/v2/...
-QUICKNODE_URL=https://....quiknode.pro/...
-
-# Plaintext RPC URLs used ONLY by `npm run test-consensus`. Lets you exercise
-# the consensus logic before you have all three provider keys. Mix paid +
-# public endpoints freely — the test harness doesn't enforce the hostname
-# whitelist. For real use the action requires three whitelisted providers.
-RPC_URL_1=
-RPC_URL_2=https://ethereum.publicnode.com
-RPC_URL_3=https://eth.drpc.org
-
-# Filled in after running `npm run encrypt-urls` ------------------------------
-ENCRYPTED_INFURA_URL=
-ENCRYPTED_ALCHEMY_URL=
-ENCRYPTED_QUICKNODE_URL=
-
-# Chain we're reading from (mainnet by default) ------------------------------
-SOURCE_CHAIN_ID=1
-
-# Contract deploy / submit ---------------------------------------------------
-DEPLOYER_PRIVATE_KEY=
-BASE_SEPOLIA_RPC_URL=https://sepolia.base.org
-BASE_RPC_URL=https://mainnet.base.org
-
-CONSENSUS_ORACLE_ADDRESS=
-REGISTRY_CHAIN_ID=84532
-REGISTRY_RPC_URL=https://sepolia.base.org
-
-# Account that pays gas to submit signed readings on-chain
-SUBMITTER_PRIVATE_KEY=
