+++ examples/multi-source-price-oracle/.env.example
@@ -1,0 +1,19 @@
+# Lit Chipotle ---------------------------------------------------------------
+LIT_API_BASE=https://api.chipotle.litprotocol.com
+LIT_USAGE_API_KEY=
+
+# Filled in by setup.js
+ACTION_WALLET_ADDRESS=
+
+# Contract deploy / submit ---------------------------------------------------
+DEPLOYER_PRIVATE_KEY=
+BASE_SEPOLIA_RPC_URL=https://sepolia.base.org
+BASE_RPC_URL=https://mainnet.base.org
+
+# Filled in by setup.js
+PRICE_ORACLE_ADDRESS=
+REGISTRY_CHAIN_ID=84532
+REGISTRY_RPC_URL=https://sepolia.base.org
+
+# Account that pays gas to submit signed readings on-chain
+SUBMITTER_PRIVATE_KEY=
