+++ examples/multi-source-price-oracle/package.json
@@ -1,17 +1,15 @@
 {
-  "name": "multi-rpc-consensus-oracle",
+  "name": "multi-source-price-oracle",
   "version": "0.0.1",
   "private": true,
-  "description": "Lit Action that reads an EVM view function from three RPC providers in parallel and signs only when all three agree.",
+  "description": "Lit Action that fetches a spot price from three independent exchanges, takes the median, and signs it for on-chain consumption.",
   "scripts": {
     "compile": "hardhat compile",
     "setup": "node scripts/setup.js",
-    "mint-pkp": "node scripts/mintPkp.js",
     "deploy:baseSepolia": "hardhat run scripts/deploy.js --network baseSepolia",
     "deploy:base": "hardhat run scripts/deploy.js --network base",
-    "encrypt-urls": "node scripts/encryptRpcUrls.js",
     "submit": "node scripts/submit.js",
-    "test-consensus": "node scripts/test-consensus.js"
+    "test-medianizer": "node scripts/test-medianizer.js"
   },
   "dependencies": {
     "@openzeppelin/contracts": "^5.0.2",
@@ -19,7 +17,7 @@
     "ethers": "^5.7.2"
   },
   "devDependencies": {
-    "@nomicfoundation/hardhat-toolbox": "^5.0.0",
+    "@nomicfoundation/hardhat-toolbox": "^2.0.2",
     "hardhat": "^2.22.10"
   }
 }
